fastapi_env/pyproject.toml
[project]
name = "fastapi-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.131.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
